Revamped matching approach. Fixed searching on "404" page.
We used to provide "Best", "Has docs" and "Everything", now we provide "Exact"
and "Substring". For "Exact", the class name must match exactly (modulo case),
for "Substring" it need only contain the query as a substring. In both cases,
if there is exactly one match that provides docs, it is shown; otherwise the
entire list of docs and non-docs matches are shown. Simpler and more useful (we
hope).